requirements.txt
celery~=5.5.2
flask~=3.1.0
flask-compress~=1.17
flask-cors~=6.0.0
flask-migrate~=4.0.7
flask-restful~=0.3.10
flask-sqlalchemy~=3.1.1
python-dotenv==1.0.1
redis[hiredis]~=6.1.0
sqlalchemy~=2.0.29
uvicorn
psycopg-binary
alembic
python-dotenv